Build 609
!! HavenBags has been updated to use some of the features from the newer version of ValorlessUtils, and therefore now require ValorlessUtils 1.5.3.173 or newer.
HavenBags will now auto-disable if a lower version is found, rather than throw errors.

You can now preview bags in the AdminGUI.

Fixed an issue causing previewed bags to return and be saved, when the server shut down or reloaded.

Fixed an issue causing sound not to play when the bag is closed.

Lang.yml value 'bag-size-error' has been removed, as it is no longer used.
Size now clamps within 1-6, so even if you put 23 the size will be capped at 6.

Added PlaceholderAPI as softdepend.

SFX are done through ValorlessUtils now, instead of internally.
Primarily done to lower the size of the plugin, as it was added to ValorlessUtils.